requirements.txt
web3>=6.0.0,<7.0.0
requests>=2.28.0
eth-abi>=4.0.0,<5.0.0
python-dotenv>=1.0.0 